conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
argon2-cffi=20.1.0=py38h27cfd23_1
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
biopython=1.78=py38h7f8727e_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bottleneck=1.3.2=py38heb32a55_1
ca-certificates=2022.4.26=h06a4308_0
certifi=2022.6.15=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
dbus=1.13.18=hb2f20db_0
debugpy=1.4.1=py38h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py38_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
glib=2.69.1=h5202010_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
hmmer=3.1b2=0
icu=58.2=he6710b0_3
importlib-metadata=4.8.1=py38h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=6.4.1=py38h06a4308_1
ipython=7.27.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.4=pyhd3eb1b0_0
jedi=0.18.0=py38h06a4308_1
jinja2=3.0.1=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py38_7
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h7f8727e_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
markupsafe=2.0.1=py38h27cfd23_0
matlabengineforpython=R2021a=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.3=py38h06a4308_0
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
openssl=1.1.1o=h7f8727e_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.3=py38h8c16a72_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=21.0.1=py38h06a4308_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pyzmq=22.2.1=py38h295c915_1
qt=5.9.7=h5867ecd_1
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
scipy=1.6.2=py38had2a1c9_1
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
terminado=0.9.4=py38h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tk=8.6.11=h1ccaba5_0
tornado=6.1=py38h27cfd23_0
traitlets=5.1.0=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py38_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
